Refactor RetinaFace pipeline and update mixed precision
Refactored BitRetinaFace to use RetinaFaceTensor for input/output, improved target handling, and modularized loss computation. Updated mixed precision training to use 'bf16' when AMP is enabled across all main scripts. Enhanced RetinaFaceTensor with normalization, tensor conversion, and deep copy utilities. Fixed minor test and hub loading issues, and updated the MobileNetV2 ternary checkpoint.

+++ BitMobileNetV2.py
@@ -242,7 +242,7 @@ def main():
         
     trainer = AccelTrainer(
         max_epochs=args.epochs,
-        mixed_precision="fp16" if args.amp else "no",
+        mixed_precision="bf16" if args.amp else "no",
         gradient_accumulation_steps=1,
         log_every_n_steps=10,
     )
