+++ BitMobileNetV4.py
@@ -702,7 +702,7 @@ def main_mnv4() -> None:
 
     trainer = AccelTrainer(
         max_epochs=args.epochs,
-        mixed_precision="fp16" if args.amp else "no",
+        mixed_precision="bf16" if args.amp else "no",
         gradient_accumulation_steps=1,
         log_every_n_steps=10,
     )
